// generated/src/app/api-models/pmb-setup.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbSetup {
  seqGpbilId: number;
  jobId: string;
  requestUser: string;
  requestDate: Date;
  requestType: string;
  action: string;
  billThruRequestDate: Date;
  billingType: string;
  billJobType: string;
  billingCycle: string;
  postDate: Date;
  template: string;
  daemonRequest: string;
  status: string;
  comments: string;
  inProcess: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  subscFlag: string;
}